requirements.txt
python-telegram-bot==13.13
pandas==2.0.2
python-dotenv~=1.0.0
aiogram~=2.25.1
asyncio~=3.4.3
nest_asyncio==1.5.7
selenium==4.11.2
apscheduler==3.6.3
tzlocal==5.0.1
datetime
openpyxl
